Turn CSS non-blocking via rel=preload (only for frontpage)
Also add a polyfill for Firefox, cause the fox doesn't support it yet. It did, but then it turned out the implementation was too buggy, so they turned it off back in Firefox 57.

Testing for frontpage. I'm not a fan of the page 'reloading' from HTML only to HTML+CSS